+++ tests/conftest.py
@@ -15,14 +15,23 @@
 
     Contact: code@inmanta.com
 """
+import glob
 import os
+import subprocess
 import sys
+import tempfile
+from pathlib import Path
+from typing import Iterator, Optional
 
 import pkg_resources
 import pytest
 
+# be careful not to import any core>=6 objects directly
+import core
 import pytest_inmanta.plugin
+from inmanta import env, loader, plugins
 from inmanta.loader import PluginModuleFinder
+from libpip2pi.commands import dir2pi
 
 pytest_plugins = ["pytester"]
 
@@ -37,13 +46,71 @@ def deactive_venv():
     old_os_path = os.environ.get("PATH", "")
     old_prefix = sys.prefix
     old_path = sys.path
+    old_pythonpath = os.environ.get("PYTHONPATH", None)
+    old_os_venv: Optional[str] = os.environ.get("VIRTUAL_ENV", None)
+    old_working_set = pkg_resources.working_set
 
     yield
 
     os.environ["PATH"] = old_os_path
     sys.prefix = old_prefix
     sys.path = old_path
-    pkg_resources.working_set = pkg_resources.WorkingSet._build_master()
-    # stay compatible with older versions of core
+    pkg_resources.working_set = old_working_set
+    # Restore PYTHONPATH
+    if old_pythonpath is not None:
+        os.environ["PYTHONPATH"] = old_pythonpath
+    elif "PYTHONPATH" in os.environ:
+        del os.environ["PYTHONPATH"]
+    # Restore VIRTUAL_ENV
+    if old_os_venv is not None:
+        os.environ["VIRTUAL_ENV"] = old_os_venv
+    elif "VIRTUAL_ENV" in os.environ:
+        del os.environ["VIRTUAL_ENV"]
+    # stay compatible with older versions of core: don't call the function if it doesn't exist
+    if hasattr(env, "mock_process_env"):
+        env.mock_process_env(python_path=sys.executable)
     if hasattr(PluginModuleFinder, "reset"):
         PluginModuleFinder.reset()
+    plugins.PluginMeta.clear()
+    loader.unload_inmanta_plugins()
+
+
+@pytest.fixture
+def examples_working_dir(pytestconfig) -> Iterator[None]:
+    examples_dir: Path = pytestconfig.rootpath / "examples"
+    current_dir: str = os.getcwd()
+    os.chdir(examples_dir)
+    yield
+    os.chdir(current_dir)
+
+
+@pytest.fixture(scope="session")
+def examples_v2_package_index(pytestconfig) -> Iterator[str]:
+    """
+    Creates a local pip index for all v2 modules in the examples dir. The modules are built and published to the index.
+
+    :return: The path to the index
+    """
+    if not core.SUPPORTS_MODULES_V2:
+        pytest.skip(
+            "Skipping modules v2 related tests for inmanta-core<6 (pre modules v2).",
+        )
+
+    examples_dir: Path = pytestconfig.rootpath / "examples"
+
+    with tempfile.TemporaryDirectory() as artifact_dir:
+        for module_dir in glob.iglob(str(examples_dir / "inmanta-module-*")):
+            subprocess.check_call(
+                [
+                    sys.executable,
+                    "-m",
+                    "inmanta.app",
+                    "module",
+                    "build",
+                    "--output-dir",
+                    artifact_dir,
+                ],
+                cwd=str(module_dir),
+            )
+        dir2pi(argv=["dir2pi", artifact_dir])
+        yield os.path.join(artifact_dir, "simple")