+++ tests/test_basic_example.py
@@ -155,3 +155,13 @@ def test_project_no_plugins(testdir):
         "DeprecationWarning: The project_no_plugins fixture is deprecated"
         " in favor of the INMANTA_TEST_NO_LOAD_PLUGINS environment variable."
     ) in result.stdout.str()
+
+
+def test_state(testdir):
+    """Make sure that importing functions works."""
+
+    testdir.copy_example("testmodule")
+
+    result = testdir.runpytest("tests/test_state.py")
+
+    result.assert_outcomes(passed=2)