+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v 2.0.0 (2022-01-24)
+Changes in this release:
+- Added support for testing v2 modules.
+- Extended to be compatible with `inmanta-core>=6`
+- Added support for custom `inmanta_reset_state` method to clean up stateful modules between compiles
+- Ensure that projects are compiled using a separate venv.
+- Fixed typing issue for `filter_args` in different method of the Project class.
+
+## Breaking changes
+- pytest-inmanta now keeps `inmanta_plugins` submodules alive across compiles. As a result, stateful modules must implement
+    custom state cleanup logic as described in the README.
+
 # v 1.6.2 (2021-08-17)
 Changes in this release:
 - Fixed issue with project fixture related to cleanup assumptions causing failures for `inmanta-core>=5.1.2.dev`